improved netupgrader

+++ netupgrade.sh
@@ -5,12 +5,14 @@ else
 	BRANCH=$1
 fi
 
-service ddos stop
-if [ -d ~/ddos-deflate-master/]; then
-	mv ~/ddos-deflate-master ~/ddos-deflate
+if [ -d "$DESTDIR/usr/local/ddos" ]; then
+	service ddos stop
+	if [ -d ~/ddos-deflate-master/]; then
+		mv ~/ddos-deflate-master ~/ddos-deflate
+	fi
+	cd ~/ddos-deflate/
+	sh uninstall.sh
+	cd ~
+	rm -rf ~/ddos-deflate/
 fi
-cd ~/ddos-deflate/
-sh uninstall.sh
-cd ~
-rm -rf ~/ddos-deflate/
 curl -s "https://raw.githubusercontent.com/phoenixweb/ddos-deflate/$BRANCH/netinstall.sh" | bash -s $BRANCH
